Rearrange options panel; implement FloorPack customisation.
