+++ Realistic Population Revisited.csproj
@@ -117,18 +117,21 @@
     <Compile Include="Code\AI_Files\AI_School.cs" />
     <Compile Include="Code\GUI\UIModCalcs.cs" />
     <Compile Include="Code\Patcher.cs" />
-    <Compile Include="Code\Settings\ConsumptionTabs\CommercialPanel.cs" />
-    <Compile Include="Code\Settings\ConsumptionTabs\ConsumptionPanelBase.cs" />
-    <Compile Include="Code\Settings\ConsumptionTabs\IndustrialPanel.cs" />
-    <Compile Include="Code\Settings\ConsumptionTabs\OfficePanel.cs" />
-    <Compile Include="Code\Settings\ConsumptionTabs\PanelBase.cs" />
-    <Compile Include="Code\Settings\ConsumptionTabs\ResidentialPanel.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\CalculationPanelBase.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\ConsumptionTabs\CommercialPanel.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\ConsumptionTabs\ConsumptionPanelBase.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\ConsumptionTabs\IndustrialPanel.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\ConsumptionTabs\OfficePanel.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\ConsumptionTabs\PanelBase.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\ConsumptionTabs\ResidentialPanel.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\FloorPanel.cs" />
     <Compile Include="Code\Settings\ModSettings.cs" />
-    <Compile Include="Code\Settings\OptionsPanelTabs\ConsumptionPanel.cs" />
-    <Compile Include="Code\Settings\OptionsPanelTabs\LegacyPanel.cs" />
+    <Compile Include="Code\Settings\OptionsPanelTabs\CalculationsPanel.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\ConsumptionPanel.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\LegacyPanel.cs" />
     <Compile Include="Code\Settings\OptionsPanelTabs\CrimePanel.cs" />
-    <Compile Include="Code\Settings\OptionsPanelTabs\DefaultsPanel.cs" />
-    <Compile Include="Code\Settings\OptionsPanelTabs\PopulationPanel.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\DefaultsPanel.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\PopulationPanel.cs" />
     <Compile Include="Code\Settings\OptionsPanelTabs\SchoolsPanel.cs" />
     <Compile Include="Code\VolumetricData\CalcData.cs" />
     <Compile Include="Code\VolumetricData\ConfigurationUtils.cs" />
@@ -151,11 +154,11 @@
     <Compile Include="Code\GUI\UIUtils.cs" />
     <Compile Include="Code\GUI\UIVolumetricPanel.cs" />
     <Compile Include="Code\Settings\OptionsKeymapping.cs" />
-    <Compile Include="Code\Settings\LegacyConsumptionTabs\LegacyCommercialPanel.cs" />
-    <Compile Include="Code\Settings\LegacyConsumptionTabs\LegacyIndustrialPanel.cs" />
-    <Compile Include="Code\Settings\LegacyConsumptionTabs\LegacyOfficePanel.cs" />
-    <Compile Include="Code\Settings\LegacyConsumptionTabs\LegacyPanelBase.cs" />
-    <Compile Include="Code\Settings\LegacyConsumptionTabs\LegacyResidentialPanel.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\LegacyConsumptionTabs\LegacyCommercialPanel.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\LegacyConsumptionTabs\LegacyIndustrialPanel.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\LegacyConsumptionTabs\LegacyOfficePanel.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\LegacyConsumptionTabs\LegacyPanelBase.cs" />
+    <Compile Include="Code\Settings\CalculationTabs\LegacyConsumptionTabs\LegacyResidentialPanel.cs" />
     <Compile Include="Code\Settings\PanelUtils.cs" />
     <Compile Include="Code\Settings\OptionsPanel.cs" />
     <Compile Include="Code\Settings\OptionsPanelTabs\ModOptionsPanel.cs" />
