+++ Translations/en.xml
@@ -153,6 +153,7 @@
     <Translation ID="RPR_OPT_PRS" String="Press any key" />
     <Translation ID="RPR_OPT_LEG" String="Legacy" />
     <Translation ID="RPR_OPT_CON" String="Consumption" />
+    <Translation ID="RPR_OPT_STO" String="Stories" />
     <Translation ID="RPR_OPT_LVL" String="Level" />
     <Translation ID="RPR_OPT_PERW" String="Amounts per worker" />
     <Translation ID="RPR_OPT_PERH" String="Amounts per household" />
